[gripper_action_controller] Add parameters to convert from joint position to opening in meters
Following the definition of control_msgs/GripperCommand, we need to be
able to define the GripperCommand in the distance between the fingers in
[m]. Currently, the controller is a pass-through position controller
which possibly is in [rad]. Using the new parameters, we can specify a
mapping to convert from the minimum/maximum opening to the
minimum/maximum position command. Unless all four new parameters are
specified, the default behaviour is retained.